\name{customDB}
\alias{customDB}
\docType{data}
\title{
Custom Gene Expression Profiles
}
\description{
Matrix containing a set of ranked gene expression profiles which can be used as custom input into classifyprofiles.
}
\usage{data(customDB)}

\details{
This is an example subset of the drugRL available in the DrugVsDiseasedata package
\pkg{DrugVsDiseasedata}, data object \code{drugRL}, based on data from the Connectivity
Map
}

\source{
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
}

\examples{
data(customDB)
## maybe str(profiles) ; plot(profiles) ...
}
\keyword{datasets}
