\name{customedge}
\alias{customedge}
\docType{data}
\title{
Edge attributes for example custom network
}
\description{
Data frame containing the required format of an edge attribute file input into Cytoscape
}
\usage{data(customedge)}

\details{
Example subset of edge attributes taken from the full reference data set of drug compounds in the cMap2data package.
}
\source{
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
}

\examples{
data(customedge)

}
\keyword{datasets}
