\name{customsif}
\alias{customsif}
\docType{data}
\title{
SIF file for custom clusters
}
\description{
Data frame containing SIF format file of a network which could be input into the classifyprofile function.
}
\usage{data(customsif)}

\details{
Example subset of edge attributes taken from the full reference data set of drug compounds in the cMap2data package.
}
\source{
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
}

\examples{
data(customsif)

}
\keyword{datasets}
